requirements.txt
pypdf
pandas
fastapi
sqlalchemy
uvicorn
python-multipart
streamlit
uv

# LangChain ecosystem
langchain
langchain-core
langchain-community
langchain-ollama
langchain-chroma
chromadb
